指摘対応（IxiaSのEnumを使用する形に修正）

+++ app/lib/persistence/db/Todo.scala
@@ -38,13 +38,13 @@ case class TodoTable[P <: JdbcProfile]()(implicit val driver: P)
     /* @3 */ def title       = column[String]         ("title",        O.Utf8Char255)
     /* @4 */ def body     = column[String]        ("body",      O.Text)
 
-    /* @5 */ def state = column[Int]("state", O.UInt64)
+    /* @5 */ def state = column[Short]("state", O.UInt64)
 
     /* @6 */ def updatedAt = column[LocalDateTime] ("updated_at", O.TsCurrent)
     /* @7 */ def createdAt = column[LocalDateTime] ("created_at", O.Ts)
 
     type TableElementTuple = (
-      Option[Id], Long, String, String, Int, LocalDateTime, LocalDateTime
+      Option[Id], Long, String, String, Short, LocalDateTime, LocalDateTime
     )
 
     // DB <=> Scala の相互のmapping定義
